+++ pe/_cy_machine.pyx
@@ -19,12 +19,14 @@ from pe._types import Memo
 from pe._grammar import Grammar
 from pe._parser import Parser
 from pe._optimize import optimize
-from pe.scanners cimport Scanner, Dot, Literal, CharacterClass, Regex
 from pe.actions import Action, Bind
 from pe.operators import Rule
 
 
 DEF FAILURE = -1
+
+
+# Parser ###############################################################
 
 cdef enum OpCode:
     FAIL = -1
@@ -80,6 +82,17 @@ cdef State* pop(State* state) except? NULL:
 
 _Binding = Tuple[str, Any]
 _Index = Dict[str, int]
+
+
+cdef class Scanner:
+    cpdef int scan(self, str s, int pos=0) except -2:
+        try:
+            return self._scan(s, pos, len(s))
+        except IndexError:
+            return FAILURE
+
+    cdef int _scan(self, str s, int pos, int slen) except -2:
+        return FAILURE
 
 
 cdef class Instruction:
@@ -305,6 +318,8 @@ cdef class _Parser:
         return state
 
 
+# Program Creation #####################################################
+
 # Captures and actions cannot be placed on these operators because of
 # their effect on the stack
 NO_CAP_OR_ACT = {CALL, COMMIT, UPDATE, RESTORE, FAILTWICE, RETURN}
@@ -493,7 +508,96 @@ def _parsing_instructions(defn):  # noqa: C901
         raise Error(f'invalid definition: {defn!r}')
 
 
-# for debugging
+# Scanners #############################################################
+
+cdef class Dot(Scanner):
+    cdef int _scan(self, str s, int pos, int slen) except -2:
+        if pos < slen:
+            return pos + 1
+        return FAILURE
+
+
+cdef class Literal(Scanner):
+    cdef str _x
+    cdef int _xlen
+
+    def __init__(self, str x):
+        self._x = x
+        self._xlen = len(x)
+
+    cdef int _scan(self, str s, int pos, int slen) except -2:
+        cdef int end = pos + self._xlen
+        if s[pos:end] != self._x:
+            return FAILURE
+        return end
+
+
+cdef class CharacterClass(Scanner):
+    cdef:
+        str _chars, _ranges
+        int _rangelen
+        bint _negate
+    cdef public:
+        int mincount, maxcount
+
+    def __init__(
+        self,
+        list ranges,
+        bint negate = False,
+        int mincount = 1,
+        int maxcount = 1
+    ):
+        self._chars = ''.join(a for a, b in ranges if not b)
+        self._ranges = ''.join(a+b for a, b in ranges if b)
+        self._rangelen = len(self._ranges)
+        self._negate = negate
+        self.mincount = mincount
+        self.maxcount = maxcount
+
+    cdef int _scan(self, str s, int pos, int slen) except -2:
+        cdef Py_UCS4 c
+        cdef str ranges = self._ranges
+        cdef bint matched
+        cdef int mincount = self.mincount
+        cdef int maxcount = self.maxcount
+        cdef int i = 0
+        while maxcount and pos < slen:
+            c = s[pos]
+            matched = False
+            if c in self._chars:
+                matched = True
+            else:
+                while i < self._rangelen:
+                    if ranges[i] <= c <= ranges[i+1]:
+                        matched = True
+                        break
+                    i += 2
+            if matched ^ self._negate:
+                pos += 1
+                maxcount -= 1
+                mincount -= 1
+            else:
+                break
+        if mincount > 0:
+            return FAILURE
+        return pos
+
+
+cdef class Regex(Scanner):
+    cdef object _regex
+
+    def __init__(self, str pattern, int flags=0):
+        self._regex = re.compile(pattern, flags=flags)
+
+    cdef int _scan(self, str s, int pos, int slen) except -2:
+        m = self._regex.match(s, pos=pos)
+        if m is None:
+            return FAILURE
+        else:
+            return m.end()
+
+
+# Debugging ############################################################
 
 # _OpCodeNames = {
 #     FAIL: 'FAIL',